+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/querycreators/GenericJsonQueryCreatorFactory.java
@@ -15,6 +15,8 @@ import org.apache.flink.table.factories.SerializationFormatFactory;
 import com.getindata.connectors.http.LookupQueryCreator;
 import com.getindata.connectors.http.LookupQueryCreatorFactory;
 import com.getindata.connectors.http.internal.table.lookup.LookupRow;
+import com.getindata.connectors.http.internal.utils.SynchronizedSerializationSchema;
+import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.ASYNC_POLLING;
 import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.LOOKUP_REQUEST_FORMAT;
 
 /**
@@ -47,8 +49,14 @@ public class GenericJsonQueryCreatorFactory implements LookupQueryCreatorFactory
             queryFormatAwareConfiguration
         );
 
-        SerializationSchema<RowData> serializationSchema =
-            encoder.createRuntimeEncoder(null, lookupRow.getLookupPhysicalRowDataType());
+        final SerializationSchema<RowData> serializationSchema;
+        if (readableConfig.get(ASYNC_POLLING)) {
+            serializationSchema = new SynchronizedSerializationSchema<>(
+                encoder.createRuntimeEncoder(null, lookupRow.getLookupPhysicalRowDataType()));
+        } else {
+            serializationSchema =
+                encoder.createRuntimeEncoder(null, lookupRow.getLookupPhysicalRowDataType());
+        }
 
         return new GenericJsonQueryCreator(serializationSchema);
     }